+++ accounting/src/sdk/billcreditnotes.ts
@@ -673,7 +673,7 @@ export class BillCreditNotes {
         billCreditNoteId: string,
         companyId: string,
         connectionId: string,
-        requestBody?: operations.UploadBillCreditNoteAttachmentRequestBody,
+        attachmentUpload?: shared.AttachmentUpload,
         retries?: utils.RetryConfig,
         config?: AxiosRequestConfig
     ): Promise<operations.UploadBillCreditNoteAttachmentResponse> {
@@ -681,7 +681,7 @@ export class BillCreditNotes {
             billCreditNoteId: billCreditNoteId,
             companyId: companyId,
             connectionId: connectionId,
-            requestBody: requestBody,
+            attachmentUpload: attachmentUpload,
         });
         const baseURL: string = utils.templateUrl(
             this.sdkConfiguration.serverURL,
@@ -696,7 +696,11 @@ export class BillCreditNotes {
         let [reqBodyHeaders, reqBody]: [object, any] = [{}, null];
 
         try {
-            [reqBodyHeaders, reqBody] = utils.serializeRequestBody(req, "requestBody", "multipart");
+            [reqBodyHeaders, reqBody] = utils.serializeRequestBody(
+                req,
+                "attachmentUpload",
+                "multipart"
+            );
         } catch (e: unknown) {
             if (e instanceof Error) {
                 throw new Error(`Error serializing request body, cause: ${e.message}`);