+++ accounting/src/sdk/directincomes.ts
@@ -894,7 +894,7 @@ export class DirectIncomes {
         companyId: string,
         connectionId: string,
         directIncomeId: string,
-        requestBody?: operations.UploadDirectIncomeAttachmentRequestBody,
+        attachmentUpload?: shared.AttachmentUpload,
         retries?: utils.RetryConfig,
         config?: AxiosRequestConfig
     ): Promise<operations.UploadDirectIncomeAttachmentResponse> {
@@ -902,7 +902,7 @@ export class DirectIncomes {
             companyId: companyId,
             connectionId: connectionId,
             directIncomeId: directIncomeId,
-            requestBody: requestBody,
+            attachmentUpload: attachmentUpload,
         });
         const baseURL: string = utils.templateUrl(
             this.sdkConfiguration.serverURL,
@@ -917,7 +917,11 @@ export class DirectIncomes {
         let [reqBodyHeaders, reqBody]: [object, any] = [{}, null];
 
         try {
-            [reqBodyHeaders, reqBody] = utils.serializeRequestBody(req, "requestBody", "multipart");
+            [reqBodyHeaders, reqBody] = utils.serializeRequestBody(
+                req,
+                "attachmentUpload",
+                "multipart"
+            );
         } catch (e: unknown) {
             if (e instanceof Error) {
                 throw new Error(`Error serializing request body, cause: ${e.message}`);