+++ accounting/RELEASES.md
@@ -59,3 +59,13 @@ Based on:
 - [typescript v0.4.3] accounting
 ### Releases
 - [NPM v0.4.3] https://www.npmjs.com/package/@speakeasy-sdks/accounting/v/0.4.3 - accounting
+
+## 2023-12-18 01:51:34
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.126.4 (2.214.10) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.5.0] accounting
+### Releases
+- [NPM v0.5.0] https://www.npmjs.com/package/@speakeasy-sdks/accounting/v/0.5.0 - accounting